Add Visualise ContentItem to publish /visualise
Following [1], the split of the start pages from smart answer flow.
The visualise tool appear to be omitted and request to this path
returned 404s/ were not found.

This commit defines a VisualiseContentItem class, the intention
here is to use objects from this to publish the /visualise path
for each smart answer to the publishing api.

At the moment, the content id is being generated each time.
it isn't clear at this time if we need to hold one consistent
content id for the visualise path. More to come on this.

[1] https://github.com/alphagov/smart-answers/pull/3126