Question 13: Replace Yes/No with list of benefits
As part of the accessibility improvements, we have been ask to replace
the Yes/No in question 13 (i.e do_either_of_the_following_apply) with
the list of benefit listed above it. The idea is to assist devices like
screen reader etc, to better recognise the benefits.

When a user selects at least on of the benefits they would be sent to
the "child benefit entitled" outcome. When none is selected and the
next steps button is triggered the user is shown the child benefit not
entitled outcome.

This commit contains implementation and tests that achives and describe
the aforementioned.

Regression test artefacts have been affected by these changes they will
be added in the next commit.